adapters/python/pyproject.toml
[project]
name = "onto-sdk"
version = "0.1.0"
description = "Python SDK for Onto Platform - Adapter Development"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nats-py>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sdk"]
